chore: make sure `fromArrayLike` is exported and move `subscribeToArray` to the `from` file
