+++ src/internal/observable/from.ts
@@ -1,4 +1,3 @@
-import { subscribeToArray } from '../util/subscribeToArray';
 import { isArrayLike } from '../util/isArrayLike';
 import { isPromise } from '../util/isPromise';
 import { Observable } from '../Observable';
@@ -134,7 +133,7 @@ function fromInteropObservable<T>(obj: any) {
  * `from` conditionals because we *know* they're dealing with an array.
  * @param array The array to emit values from
  */
-function fromArrayLike<T>(array: ArrayLike<T>) {
+export function fromArrayLike<T>(array: ArrayLike<T>) {
   return new Observable((subscriber: Subscriber<T>) => {
     subscribeToArray(array, subscriber);
   });
@@ -186,6 +185,28 @@ async function process<T>(asyncIterable: AsyncIterable<T>, subscriber: Subscribe
     if (subscriber.closed) {
       return;
     }
+  }
+  subscriber.complete();
+}
+
+/**
+ * Subscribes to an ArrayLike with a subscriber
+ * @param array The array or array-like to subscribe to
+ * @param subscriber
+ */
+export function subscribeToArray<T>(array: ArrayLike<T>, subscriber: Subscriber<T>) {
+  // Loop over the array and emit each value. Note two things here:
+  // 1. We're making sure that the subscriber is not closed on each loop.
+  //    This is so we don't continue looping over a very large array after
+  //    something like a `take`, `takeWhile`, or other synchronous unsubscription
+  //    has already unsubscribed.
+  // 2. In this form, reentrant code can alter that array we're looping over.
+  //    This is a known issue, but considered an edge case. The alternative would
+  //    be to copy the array before executing the loop, but this has
+  //    performance implications.
+  const length = array.length;
+  for (let i = 0; i < length && !subscriber.closed; i++) {
+    subscriber.next(array[i]);
   }
   subscriber.complete();
 }
